pyproject.toml
[tool.poetry]
name = "builder"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src", from = "Builder" }]

[tool.poetry.dependencies]
python = "^3.11"
argparse = "^1.4.0"
attrs = "^23.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
